refactor: 💡 cleanup global dateHelper functions
